Move subscribeToPrivateUserChannelEvent to Pusher
